GameRunner - merge TripleAProcessRunner into GameRunner (delete TripleAProcessRunner)
