+++ src/games/strategy/engine/lobby/client/ui/LobbyGamePanel.java
@@ -26,7 +26,8 @@ import javax.swing.ListSelectionModel;
 import javax.swing.SwingUtilities;
 import javax.swing.table.DefaultTableCellRenderer;
 
-import games.strategy.engine.framework.TripleAProcessRunner;
+import games.strategy.ui.SwingAction;
+import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.startup.ui.InGameLobbyWatcher;
 import games.strategy.engine.framework.startup.ui.ServerOptions;
 import games.strategy.engine.lobby.server.AbstractModeratorController;
@@ -35,7 +36,6 @@ import games.strategy.engine.lobby.server.IModeratorController;
 import games.strategy.net.INode;
 import games.strategy.net.Messengers;
 import games.strategy.net.Node;
-import games.strategy.ui.SwingAction;
 import games.strategy.util.MD5Crypt;
 
 public class LobbyGamePanel extends JPanel {
@@ -309,7 +309,7 @@ public class LobbyGamePanel extends JPanel {
     // we sort the table, so get the correct index
     final int modelIndex = m_tableSorter.modelIndex(selectedIndex);
     final GameDescription description = m_gameTableModel.get(modelIndex);
-    TripleAProcessRunner.joinGame(description, m_messengers, getParent());
+    GameRunner.joinGame(description, m_messengers, getParent());
   }
 
   protected void hostGame() {
@@ -321,7 +321,7 @@ public class LobbyGamePanel extends JPanel {
     if (!options.getOKPressed()) {
       return;
     }
-    TripleAProcessRunner.hostGame(options.getPort(), options.getName(), options.getComments(), options.getPassword(),
+    GameRunner.hostGame(options.getPort(), options.getName(), options.getComments(), options.getPassword(),
         m_messengers);
   }
